refactor: rename cst_lowering module to just cst
